fundseeker/requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
typer>=0.9.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
